Fix some TODOs and pull out shared_state.rs
