+++ wasabi/src/main.rs
@@ -45,6 +45,7 @@ mod bytes;
 mod go;
 mod js;
 mod network;
+mod shared_state;
 mod timeout_heap;
 mod util;
